[improvement][auth] Optimize the TokenService-related code
